Remove Movie.list class variable
The Movie class also lost Movie.reset and the Movie.list reader method.
I could have made the Scrape_list method only run if Movie.list was empty but ultimately that instance variable was no longer necessary as a seperate local variable is what the Cli is working with anywayso its original caching function is gone.
Removed the list variable from Scraper.scrape_list as it was useless trying to control the return value.
Finally the last commit was an accidental duplicate