Readded caching of the movie list
Added a LIST constant to Movie class and a self.all method to access it.
Modified the self.create_list method of the Movie class to save to the LIST constant instead of a local array.
Added a conditional to the Scraper classes self.scrape_list method so that it only scrapes when the Movie.all method returns an empty array, and it now returns the list through the Movie.all method.